hf_space/requirements.txt
transformers>=4.50.0
torch>=2.3.0
gradio>=4.40.0
spaces
Pillow>=10.0.0
accelerate>=0.30.0
